import from repository in side_test and plot_top_n, including bug fixes in import_from_repository that handles StimInfo tables with a subset of stimIds of experiment.

+++ EStimShapeAnalysis/src/analysis/ga/cached_ga_fields.py
@@ -40,12 +40,12 @@ class GAResponseField(StimSpecIdField):
 
 
 class ParentIdField(StimSpecIdField):
-    def get(self, task_id) -> float:
+    def get(self, task_id) -> int:
         stim_spec_id = self.get_cached_super(task_id, StimSpecIdField)
         self.conn.execute("SELECT parent_id FROM StimGaInfo WHERE stim_id = %s",
                           params=(stim_spec_id,))
         ga_response = self.conn.fetch_all()
-        return float(ga_response[0][0])
+        return int(ga_response[0][0])
 
     def get_name(self):
         return "ParentId"
